+++ src/components/layout/RootLayout.tsx
@@ -3,6 +3,7 @@ import { Geist, Geist_Mono } from "next/font/google";
 import "../../app/globals.css";
 import React from "react";
 import AuthProvider from "@/components/auth/AuthProvider";
+import Link from "next/link";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -29,6 +30,17 @@ export default function RootLayout({ children }: RootLayoutProps) {
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
+        <header className="bg-white shadow">
+          <div className="mx-auto max-w-7xl px-4 py-4 sm:px-6 lg:px-8">
+            <div className="flex items-center justify-between">
+              <div className="flex items-center">
+                <Link href="/" className="text-xl font-bold text-gray-900">
+                  It.dev
+                </Link>
+              </div>
+            </div>
+          </div>
+        </header>
         <AuthProvider>{children}</AuthProvider>
       </body>
     </html>
